no summary cost centres for motivation details
This has been a dirty workaround, to have a motivation detail with a summary cost centre, and then to select a reporting cost centre on the bank import match tool.
This needs to be cleaned up.
The user has to create a motivation detail for each reporting costcentre, which is the same behaviour as in Petra 2.x.
The gift entry screens are already complex enough, supporting summary cost centres would be even more complicated.