Fix decoding and encoding of composite variants and empty strings.
Originally, a Variant created with an empty string would be encoded into
a string Variant composed of a single space character. This is fixed by
using the CSV encoding function properly. Also, the composite Variants
were encoded and decoded in ambiguous ways. Now the fields in composite
Variants are properly quoted and escaped when encoded.

The test cases are updated to reflect the new behavior. The old test
cases approved the old misbehavior of an empty string Variant morphing
into a single-space string Variant, but this is fixed as well.