New branch from trunk - augmented with the changes of Seth that were contained in Branch 'dev_0000354_studentsJ2011_usercontrol_tasklist_refinement'. We abandoned that Branch as it was so old and merging trunk to it gave problems because of that.
Note #1: The pnlCollapsible and TaskList Controls don't work properly in the ControlTestBench. However, that is the state they were left in.

Note #2: MainForm3 is the Main Form that Seth last launched, and that is what is now opened when ControlTestBench is run. MainForm is the original test bench form and left in this Branch for reference, at least for the moment.