In order to get the tests to work for Daily Exchange Rate I had to revert the server DataReader code to what it was originally.  This now calls LoadAll on the DailyExchangeRate table.  To compensate for that I have modified the WindowMaintainTable template so that if there is a DataSet defined it does not do an initail load and Merge, but expects the manual code to do this using an appropriate call.  This is fair enough since this is the only screen that has a dataSet defined in a WindowmaintainTable.
I then had to make a few changes to the test suite.  The main thing I have done in that regard though is to disable the tests that deal with exchange rate usage because the screen works a bit differently.  I will need to do more work on that as well as create some new tests to run on the server side.  However I want to check the code in and start a new branch for that.